pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copy_trader"
version = "0.1.0"
description = "Standalone copy-trading program: Hyperliquid → Binance perpetual position sync"
requires-python = ">=3.13,<3.14"
dependencies = [
    # HTTP client
    "httpx>=0.27",
    # YAML config loading
    "PyYAML>=6.0",
    # .env loading
    "python-dotenv>=1.0",
    # Data validation / config models
    "pydantic>=2.6",
    # Binance REST API
    "binance-futures-connector>=4.0",
    # Precise decimal arithmetic  (already stdlib, but kept explicit)
    # stdlib: decimal, sqlite3, dataclasses, asyncio
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",        # httpx mock for tests
    "ruff>=0.4",
]

[project.scripts]
copy-trader = "copy_trader.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
